Ios 781

+++ WavesSDK/Sources/Signature/TransactionSignatureV1.swift
@@ -140,7 +140,7 @@ public extension TransactionSignatureV1 {
             public struct Arg {
                 public enum Value {
                     case bool(Bool) //boolean
-                    case integer(Int) // integer
+                    case integer(Int64) // integer
                     case string(String) // string
                     case binary(String) // binary
                 }
